Revamp webmock from one giant file to move logic closer to specfiic tests
